Prevent scan from interfering with combine/sample.
This is a proof-of-concept solution to the problem of scan causing
delayed observation of the incoming stream.  It was using cons, which
uses concat, which uses continueWith, and continueWith delays
observation.  If the previous stream was state-dependent, like combine
or sample, they might miss an event, and thus never transition from
waiting to ready state.

It may be better to "fix" this by changing cons (aka startWith), or
invent a new operation that observes two streams, A and B,
immediately, and only propagates events from A until it ends, and then
starts propagating from B.